raspberry/requirements.txt
pymavlink>=2.4.30
pyserial>=3.5
